Add products model and routes
Note : Delete doesn't work, due to an sql error. When done directly in SQL request, I've manage to delete the error permenantly, but I'm completly lock via sequelize, I don't understand why
The error is that when a product is delete, a trigger is, well, triggered, and try to put the product data in the trashed databased. Sequelize insist that "IDENTITY_INSERT is set to OFF" when I turned it on and it worked afterward in sql requests
If i don't find a solution in the long run, I might just delete the trigger.
For now, I'll just do something else